requirements.txt
pandas
requests
beautifulsoup4
google-generativeai
python-dotenv
